Flatten supervisor tree
There is no need in separate `Block.Supervisor` hierarchy, so its children
are placed in the root supervisor tree.

The tree itself is moved to `Indexer.Supervisor` module to reduce amount
of directories.
